Update dao_macro.c
Possible fix for several syntax-related crashes as well as for `use syntax mylang` not functioning correctly.

Crashes were caused by infinite recursion in DaoParser_ErrorX2 and DaoParser_ErrorX3. Use statement breakage was due to not seeking ahead far enough in the token list after parsing the use statement. This may not be the correct fix for that; there may be a need to actually do some parsing to figure out how many tokens forward to seek. But this worked in my tests.